Split the log out of the console as RunLog, and called it Log in the UI
The two views of a run were named a level apart: the log was
Console.ListView, a property, while the canvas was a top-level file with a
namespace of its own. Pushing the log out beats pulling the canvas in — the
canvas is a first-class concept a script draws on, and burying it to fix a
symmetry problem would be the wrong trade. Console is now purely the frame:
the run picker, the view switch, and whichever view is showing.

The parts moved with it, which is the point rather than the rename:
RunLog.CallRow / TailBar / NoPayload / PayloadPane / PayloadTabs /
ResponseSummary / HeadersContent / HeadersTable belong to the log,
Console.RunSelect and Console.RunRow to the frame. Flat against one another
they read as the same kind of thing. The virtualisation constants went with
the rows they measure; only the frame still reads the consoles store.

"RunLog" over the alternatives: Log alone collides with the script's own log
messages (Log from server/api, ConsoleItem.logs, Canvas.Logs), and RunList
would be read as a list of runs — which is exactly what the run picker
beside it is.

The segmented control now says Log and Canvas rather than List and Canvas,
since the log is what the docs have always called it. ConsoleView is
"log" | "canvas" to match, and it is in-memory only, so nothing had to
migrate.
