+++ ui/src/Console.tsx
@@ -1,7 +1,7 @@
-import { ArrowDown, Bot, Check, ChevronDown, ChevronsUpDown, ChevronUp, Copy, FoldVertical, Trash2, UnfoldVertical } from "lucide-react";
-import { Fragment, memo, useCallback, useEffect, useLayoutEffect, useRef, useState, useSyncExternalStore } from "react";
+import { Bot, Check, ChevronDown, ChevronsUpDown, ChevronUp, Copy, FoldVertical, Trash2, UnfoldVertical } from "lucide-react";
+import { Fragment, memo, useCallback, useEffect, useRef, useState, useSyncExternalStore } from "react";
 import { ApproveGesture, blockText } from "./blocks";
-import { barFraction, callErrorCode, dotClass, formatBytes, formatDuration, payloadBytes, statusClass } from "./callFormat";
+import { dotClass, formatDuration } from "./callFormat";
 import { formatClockTime, formatDayLabel, formatElapsed, isSameDay } from "./callTime";
 import { Canvas } from "./Canvas";
 import { cn } from "./cn";
@@ -10,14 +10,11 @@ import { IconButton } from "./components/icon-button";
 import { SegmentedControl } from "./components/segmented-control";
 import { Spinner } from "./components/spinner";
 import { consoles } from "./consoles";
-import { unwrapEnvelope } from "./httpEnvelope";
-import { JsonViewer, JsonViewerHandle } from "./JsonViewer";
-import { MethodCall } from "./kaja";
-import { callStatus, ConsoleItem, ConsoleTab, ConsoleView, defaultView, followSelection, itemStatus, RunGroup, RunSelection, RunStatus } from "./runs";
+import { JsonViewerHandle } from "./JsonViewer";
+import { RunLog } from "./RunLog";
+import { ConsoleItem, ConsoleTab, ConsoleView, defaultView, followSelection, RunGroup, RunSelection } from "./runs";
 import { runShortcutLabel } from "./RunButton";
 import { TableView } from "./tableView";
-import { LogLevel } from "./server/api";
-import { unwrapFailure, upstreamRequestLine } from "./upstreamHeaders";
 
 export type { ConsoleItem } from "./runs";
 
@@ -25,29 +22,10 @@ export type { ConsoleItem } from "./runs";
 // never turns the console header into a full-height surface.
 const RUN_ROW_HEIGHT = 32;
 const MAX_VISIBLE_RUN_ROWS = 8;
-// The log's rows are a fixed height, which is what makes windowing it arithmetic
-// rather than measurement: the row that belongs at a scroll position is a
-// division, and the ones above and below it are two spacers.
-const CALL_ROW_HEIGHT = 24;
-// Rows drawn beyond the viewport, so a flick of the wheel doesn't show a gap.
-const OVERSCAN = 8;
-// How close to the bottom still counts as being at the bottom. A run appends
-// while you read, so this is a couple of rows rather than nothing.
-const TAIL_SLACK = CALL_ROW_HEIGHT * 2;
-// How much of the console the log may take before the payload pane stops
-// shrinking. The pane is the thing being read; the log is how you choose it.
-const MAX_LOG_HEIGHT = "45%";
-// The widest a duration bar is drawn, in pixels.
-const BAR_WIDTH = 88;
 
-const consoleTabClass = "cursor-pointer select-none whitespace-nowrap text-xs text-muted-foreground hover:text-foreground";
-const consoleTabActiveClass = "font-medium text-foreground";
 // Utilities carry no resting chrome: they are worth no more weight than the tabs
 // beside them.
 const utilityButtonClass = "h-6 w-6 rounded-md hover:bg-accent hover:text-foreground";
-// Both time columns are reserved: the longest value they can ever hold sets the
-// geometry once, so nothing moves as calls settle and age.
-const DURATION_COLUMN_CLASS = "w-[9ch] shrink-0 truncate text-right font-mono text-xs tabular-nums text-muted-foreground";
 
 interface ConsoleProps {
   // Which file's console this is. It is the whole scope: the runs are that
@@ -193,7 +171,7 @@ export function Console({ fileId, onAnswer, onCancelAsk, onDecide, tableViews, o
     (itemId: string) => {
       if (!selectedGroup) return;
       setTailing(false);
-      onViewChange("list");
+      onViewChange("log");
       onSelect({ runId: selectedGroup.run.id, itemId });
     },
     [selectedGroup, onSelect, onViewChange],
@@ -236,12 +214,12 @@ export function Console({ fileId, onAnswer, onCancelAsk, onDecide, tableViews, o
         <div className="h-4 w-px shrink-0 bg-border" />
         <SegmentedControl className="h-[26px] shrink-0 p-[2px]" aria-label="Run view">
           <SegmentedControl.Button
-            selected={activeView === "list"}
+            selected={activeView === "log"}
             className="h-[20px] px-2.5 py-0 text-xs"
-            onClick={() => onViewChange("list")}
-            data-testid="console-view-list"
+            onClick={() => onViewChange("log")}
+            data-testid="console-view-log"
           >
-            List
+            Log
           </SegmentedControl.Button>
           <SegmentedControl.Button
             selected={activeView === "canvas"}
@@ -282,7 +260,7 @@ export function Console({ fileId, onAnswer, onCancelAsk, onDecide, tableViews, o
         </div>
         {showUtilities && (
           <div className="ml-auto flex shrink-0 items-center gap-2 @max-[430px]:hidden">
-            {activeView === "list" && (
+            {activeView === "log" && (
               <>
                 <IconButton
                   icon={FoldVertical}
@@ -328,7 +306,7 @@ export function Console({ fileId, onAnswer, onCancelAsk, onDecide, tableViews, o
           onTablePull={onTablePull}
         />
       ) : selectedGroup ? (
-        <Console.ListView
+        <RunLog
           group={selectedGroup}
           rows={rows}
           selectedItemId={selection?.itemId}
@@ -348,308 +326,6 @@ export function Console({ fileId, onAnswer, onCancelAsk, onDecide, tableViews, o
     </div>
   );
 }
-
-interface ListViewProps {
-  group: RunGroup;
-  rows: ConsoleItem[];
-  selectedItemId?: string;
-  activeTab: ConsoleTab;
-  selectedItem?: ConsoleItem;
-  waiting: boolean;
-  jsonViewerRef: React.MutableRefObject<JsonViewerHandle | null>;
-  now: number;
-  tailing: boolean;
-  // The same answer, readable without a render — see the note where it is made.
-  tailingRef: React.MutableRefObject<boolean>;
-  onTailingChange: (tailing: boolean) => void;
-  onSelectRow: (itemId: string) => void;
-  onTabChange: (tab: ConsoleTab) => void;
-  onGoToCanvas: () => void;
-}
-
-/**
- * The flat audit log. A row is a call and only a call — no disclosure triangles,
- * no block rows, no run row — which is what keeps it scannable and lets every
- * row carry the same two extra channels.
- *
- * Only the rows on screen are drawn. The rest are two spacers, because a fixed
- * row height means the log's length is a number rather than a measurement — so
- * it stays complete at any length without a thousand rows in the document.
- */
-Console.ListView = function ({
-  group,
-  rows,
-  selectedItemId,
-  activeTab,
-  selectedItem,
-  waiting,
-  jsonViewerRef,
-  now,
-  tailing,
-  tailingRef,
-  onTailingChange,
-  onSelectRow,
-  onTabChange,
-  onGoToCanvas,
-}: ListViewProps) {
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const [window, setWindow] = useState({ top: 0, height: 0 });
-  const onTailingRef = useRef(onTailingChange);
-  onTailingRef.current = onTailingChange;
-
-  const total = rows.length;
-  const slowest = group.stats.slowest;
-  const failures = group.failures;
-  const scriptFailed = group.items.some((item) => item.logs?.some((log) => log.level === LogLevel.LEVEL_ERROR));
-
-  useLayoutEffect(() => {
-    const element = scrollRef.current;
-    if (!element) return;
-
-    const measure = () => {
-      setWindow({ top: element.scrollTop, height: element.clientHeight });
-      const atBottom = element.scrollHeight - element.scrollTop - element.clientHeight <= TAIL_SLACK;
-      // Written now, not on the render this schedules: the next row may land
-      // before that render does, and it must not scroll the log back down.
-      if (atBottom !== tailingRef.current) onTailingRef.current(atBottom);
-    };
-
-    measure();
-    element.addEventListener("scroll", measure, { passive: true });
-    const observer = new ResizeObserver(measure);
-    observer.observe(element);
-    return () => {
-      element.removeEventListener("scroll", measure);
-      observer.disconnect();
-    };
-  }, [tailingRef]);
-
-  // Following means staying at the bottom as rows arrive.
-  useLayoutEffect(() => {
-    const element = scrollRef.current;
-    if (element && tailingRef.current) element.scrollTop = element.scrollHeight;
-  }, [total, tailing, tailingRef]);
-
-  const first = Math.max(0, Math.floor(window.top / CALL_ROW_HEIGHT) - OVERSCAN);
-  const count = Math.max(0, Math.min(total - first, Math.ceil(window.height / CALL_ROW_HEIGHT) + OVERSCAN * 2));
-  const visible = rows.slice(first, first + count);
-  // What is still below the fold. The log is never collapsed or summarised away,
-  // so this says what is out of sight rather than standing in for it.
-  const rowsBelow = Math.max(0, total - Math.round((window.top + window.height) / CALL_ROW_HEIGHT));
-
-  return (
-    <div className="flex min-h-0 flex-1 flex-col">
-      <div ref={scrollRef} data-testid="console-log" className="@container shrink overflow-y-auto" style={{ maxHeight: MAX_LOG_HEIGHT }}>
-        {total === 0 ? (
-          <div className="flex h-[24px] items-center px-3 text-xs text-muted-foreground">
-            {/* A run parked on a question is in flight but is not on its way to
-                a call — the tail bar below already says what it is doing. */}
-            {group.inFlight && !waiting ? "Waiting for the first call…" : "No calls."}
-          </div>
-        ) : (
-          <div style={{ height: total * CALL_ROW_HEIGHT }}>
-            <div style={{ transform: `translateY(${first * CALL_ROW_HEIGHT}px)` }}>
-              {visible.map((item) => (
-                <Console.CallRow
-                  key={item.id}
-                  id={item.id}
-                  name={`${item.call!.service.name}.${item.call!.method.name}`}
-                  timestamp={item.timestamp}
-                  loopKey={item.key}
-                  status={itemStatus(item)}
-                  durationMs={item.call!.durationMs}
-                  errorCode={callErrorCode(item.call!)}
-                  fraction={barFraction(item.call!.durationMs, slowest)}
-                  selected={item.id === selectedItemId}
-                  stale={group.run.stale === true}
-                  onSelect={onSelectRow}
-                  now={now}
-                />
-              ))}
-            </div>
-          </div>
-        )}
-      </div>
-
-      <Console.TailBar
-        waiting={waiting}
-        scriptFailed={scriptFailed}
-        rowsBelow={rowsBelow}
-        failures={failures}
-        dropped={group.dropped}
-        tailing={tailing}
-        onFollow={() => onTailingChange(true)}
-        onGoToCanvas={onGoToCanvas}
-      />
-
-      {/* The payload sits in a pane of its own that never reflows as you move
-          through the log — which is why Request/Response/Headers live down here
-          rather than in the header. */}
-      <div className={cn("flex min-h-0 flex-1 flex-col border-t border-border", group.run.stale && "opacity-70")}>
-        {group.run.payloadsExpired ? (
-          <Console.NoPayload>Response no longer kept — run to see it live</Console.NoPayload>
-        ) : selectedItem?.payloadsDropped ? (
-          <Console.NoPayload>Payload let go to keep this run bounded — run to see it live</Console.NoPayload>
-        ) : selectedItem?.call ? (
-          <Console.PayloadPane methodCall={selectedItem.call} activeTab={activeTab} onTabChange={onTabChange} jsonViewerRef={jsonViewerRef} />
-        ) : (
-          <div className="flex min-h-0 flex-1 items-center justify-center text-xs text-muted-foreground">
-            {total === 0 ? "Nothing to show." : "Select a call."}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-// A payload that is not there any more, and why. Expiry is only bearable when it
-// is a stated state rather than a silent hole.
-Console.NoPayload = function ({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="flex items-center gap-2 px-4 py-3">
-      <span className="text-xs text-muted-foreground">{children}</span>
-      <span className="font-mono text-xs text-muted-foreground">{runShortcutLabel}</span>
-    </div>
-  );
-};
-
-interface TailBarProps {
-  waiting: boolean;
-  scriptFailed: boolean;
-  rowsBelow: number;
-  failures: number;
-  // Rows a very long run stopped keeping. The log says where it stops being
-  // complete rather than quietly ending there.
-  dropped: number;
-  tailing: boolean;
-  onFollow: () => void;
-  onGoToCanvas: () => void;
-}
-
-/**
- * What the log can't say inside a row. A run parked on a question, or one whose
- * script threw, is a fact about the whole run, and the log stays readable while
- * it waits — the pause blocks the script, not your reading.
- */
-Console.TailBar = function ({ waiting, scriptFailed, rowsBelow, failures, dropped, tailing, onFollow, onGoToCanvas }: TailBarProps) {
-  if (!waiting && !scriptFailed && rowsBelow <= 0 && failures === 0 && dropped === 0 && tailing) return null;
-
-  const state = waiting ? "waiting" : scriptFailed ? "failed" : "counts";
-
-  return (
-    <div
-      data-testid="console-tail"
-      className={cn(
-        "flex h-[26px] shrink-0 items-center gap-2 border-t px-3 font-mono text-xs",
-        state === "waiting" && "border-l-2 border-l-amber-500 border-t-border bg-amber-500/10",
-        state === "failed" && "border-l-2 border-l-destructive border-t-border bg-destructive/10",
-        state === "counts" && "border-t-border",
-      )}
-    >
-      {state === "waiting" && (
-        <>
-          <span className="h-1.5 w-1.5 shrink-0 rounded-full bg-amber-500" />
-          <span className="text-amber-600 dark:text-amber-400">Waiting for an answer</span>
-        </>
-      )}
-      {state === "failed" && <span className="text-destructive">Script failed</span>}
-      {state === "counts" && rowsBelow > 0 && <span className="text-muted-foreground">{rowsBelow} more</span>}
-      <div className="ml-auto flex shrink-0 items-center gap-3">
-        {dropped > 0 && <span className="text-muted-foreground">{dropped} not kept</span>}
-        {failures > 0 && <span className="text-amber-600 dark:text-amber-400">{failures === 1 ? "1 error" : `${failures} errors`}</span>}
-        {!tailing && (
-          <button
-            type="button"
-            data-testid="console-follow"
-            className="flex items-center gap-1 rounded-full border border-border px-2 py-0.5 text-xs text-muted-foreground hover:text-foreground"
-            onClick={onFollow}
-          >
-            <ArrowDown size={11} />
-            Latest
-          </button>
-        )}
-        {(state === "waiting" || state === "failed") && (
-          <button
-            type="button"
-            className={cn(
-              "rounded-full border px-2 py-0.5 text-xs",
-              state === "waiting" ? "border-amber-500 text-amber-600 dark:text-amber-400" : "border-destructive text-destructive",
-            )}
-            onClick={onGoToCanvas}
-          >
-            Go to canvas
-          </button>
-        )}
-      </div>
-    </div>
-  );
-};
-
-interface CallRowProps {
-  id: string;
-  name: string;
-  timestamp: number;
-  loopKey?: string;
-  status: RunStatus;
-  durationMs?: number;
-  errorCode?: string;
-  fraction?: number;
-  selected: boolean;
-  stale: boolean;
-  onSelect: (itemId: string) => void;
-  now: number;
-}
-
-/**
- * One call, and the same shape for every one of them: when it happened, what it
- * was, which pass through the loop it belonged to, and how long it took — as a
- * number and as a length.
- *
- * Every prop is a value rather than an object, which is what makes the memo
- * hold: a settled row is handed the same twelve values on every repaint and
- * doesn't render again. `now` is the exception and is passed as zero unless the
- * row is the one counting up.
- */
-Console.CallRow = memo(function CallRow({
-  id,
-  name,
-  timestamp,
-  loopKey,
-  status,
-  durationMs,
-  errorCode,
-  fraction,
-  selected,
-  stale,
-  onSelect,
-  now,
-}: CallRowProps) {
-  const pending = status === "pending" || status === "streaming";
-
-  return (
-    <div
-      data-testid="console-call-row"
-      className={cn("flex shrink-0 cursor-pointer items-center gap-2.5 px-3", selected ? "bg-accent" : "hover:bg-accent/50", stale && "opacity-75")}
-      style={{ height: CALL_ROW_HEIGHT }}
-      onClick={() => onSelect(id)}
-    >
-      {pending ? <Spinner className="size-3" /> : <span className={cn("h-1.5 w-1.5 shrink-0 rounded-full", dotClass(status))} />}
-      <span className="w-[8ch] shrink-0 font-mono text-xs tabular-nums text-muted-foreground @max-[360px]:hidden">{formatClockTime(timestamp)}</span>
-      <span className={cn("min-w-0 flex-1 truncate font-mono text-xs", selected ? "text-foreground" : "text-muted-foreground")}>{name}</span>
-      {loopKey && <span className="shrink-0 truncate font-mono text-xs text-muted-foreground/80 @max-[440px]:hidden">{loopKey}</span>}
-      {errorCode && <span className="shrink-0 font-mono text-xs text-destructive">{errorCode}</span>}
-      {fraction !== undefined && (
-        <span className="shrink-0 @max-[500px]:hidden" style={{ width: BAR_WIDTH }} aria-hidden>
-          <span
-            className={cn("block h-[6px] rounded-sm", status === "error" ? "bg-destructive/40" : "bg-muted-foreground/30")}
-            style={{ width: `${fraction * 100}%` }}
-          />
-        </span>
-      )}
-      <span className={DURATION_COLUMN_CLASS}>{pending ? formatElapsed(now - timestamp) : formatDuration(durationMs)}</span>
-    </div>
-  );
-});
 
 interface RunSelectProps {
   groups: RunGroup[];
@@ -754,203 +430,6 @@ Console.RunRow = memo(function RunRow({ name, detail, dotClass: dot, pending, wa
     </DropdownMenuItem>
   );
 });
-
-interface PayloadPaneProps {
-  methodCall: MethodCall;
-  activeTab: ConsoleTab;
-  onTabChange: (tab: ConsoleTab) => void;
-  jsonViewerRef: React.MutableRefObject<JsonViewerHandle | null>;
-}
-
-Console.PayloadPane = function ({ methodCall, activeTab, onTabChange, jsonViewerRef }: PayloadPaneProps) {
-  const isStreaming = methodCall.streamOutputs !== undefined;
-  const hasResponse = methodCall.output !== undefined || methodCall.error !== undefined || (isStreaming && methodCall.streamOutputs!.length > 0);
-  const hasError = methodCall.error !== undefined;
-
-  // Switch to response tab when response arrives
-  useEffect(() => {
-    if (hasResponse && activeTab === "request") {
-      onTabChange("response");
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [hasResponse]);
-
-  let content;
-  let rawText: string | undefined;
-  if (activeTab === "request") {
-    content = methodCall.input;
-  } else if (hasError) {
-    // Same rule as the response below: an HTTP failure arrives wrapped in what
-    // carried it here, and the body the API sent is the failure.
-    content = unwrapFailure(methodCall.error);
-  } else if (isStreaming) {
-    rawText = methodCall.streamOutputs!.map((msg) => JSON.stringify(unwrapEnvelope(methodCall.outputType, msg), null, 2)).join("\n\n");
-  } else {
-    // An app that carries HTTP inside gRPC has to put a body protobuf has no
-    // shape for — an array, a scalar — in a field of its own. That field is the
-    // encoding, not the response, so the response is what it holds.
-    content = unwrapEnvelope(methodCall.outputType, methodCall.output);
-  }
-
-  return (
-    <div className="flex min-h-0 flex-1 flex-col">
-      {/* The pane names itself: which part of the call on the left, and what
-          came back of it on the right. Without the readout a successful call and
-          an empty one look the same. */}
-      <div className="flex h-[28px] shrink-0 items-center gap-4 overflow-hidden px-3">
-        <Console.PayloadTabs methodCall={methodCall} activeTab={activeTab} onTabChange={onTabChange} />
-        {activeTab !== "headers" && <Console.ResponseSummary methodCall={methodCall} content={content} rawText={rawText} />}
-      </div>
-      {activeTab === "headers" ? (
-        <Console.HeadersContent methodCall={methodCall} />
-      ) : activeTab === "response" && !hasResponse ? (
-        <div className="flex flex-1 items-center justify-center text-xs text-muted-foreground">Waiting for a response…</div>
-      ) : (
-        <>
-          {activeTab === "response" && hasError && methodCall.url && (
-            <div className="border-y border-border bg-destructive/10 px-4 py-1.5 font-mono text-xs text-destructive">POST {methodCall.url}</div>
-          )}
-          <JsonViewer ref={jsonViewerRef} value={content} rawText={rawText} />
-        </>
-      )}
-    </div>
-  );
-};
-
-interface PayloadTabsProps {
-  methodCall: MethodCall;
-  activeTab: ConsoleTab;
-  onTabChange: (tab: ConsoleTab) => void;
-}
-
-Console.PayloadTabs = function ({ methodCall, activeTab, onTabChange }: PayloadTabsProps) {
-  const isStreaming = methodCall.streamOutputs !== undefined;
-  const streamCount = isStreaming ? methodCall.streamOutputs!.length : 0;
-
-  const tab = (id: ConsoleTab, label: string) => (
-    <span className={cn(consoleTabClass, activeTab === id && consoleTabActiveClass)} onClick={() => onTabChange(id)}>
-      {label}
-    </span>
-  );
-
-  return (
-    <div className="flex shrink-0 items-center gap-4">
-      {tab("request", "Request")}
-      {tab("response", `Response${isStreaming && streamCount > 0 ? ` (${streamCount})` : ""}`)}
-      {tab("headers", "Headers")}
-    </div>
-  );
-};
-
-interface ResponseSummaryProps {
-  methodCall: MethodCall;
-  content: unknown;
-  rawText?: string;
-}
-
-// What came back, how long it took and how big it is. Status colour appears
-// here and in the call's dot, and nowhere else.
-Console.ResponseSummary = function ({ methodCall, content, rawText }: ResponseSummaryProps) {
-  const status = callStatus(methodCall);
-  const label = { pending: "Pending", streaming: "Streaming", success: "OK", error: callErrorCode(methodCall) ?? "Error" }[status];
-  const duration = formatDuration(methodCall.durationMs);
-  const size = formatBytes(payloadBytes(content, rawText));
-  const streamCount = methodCall.streamOutputs?.length;
-
-  return (
-    <div className="ml-auto flex shrink-0 items-center gap-3 overflow-hidden whitespace-nowrap font-mono text-xs">
-      <span data-testid="console-status" className={cn("shrink-0 font-medium", statusClass(status))}>
-        {label}
-      </span>
-      {duration && <span className="shrink-0 tabular-nums text-muted-foreground @max-[430px]:hidden">{duration}</span>}
-      {size && <span className="shrink-0 tabular-nums text-muted-foreground @max-[500px]:hidden">{size}</span>}
-      {streamCount !== undefined && (
-        <span className="shrink-0 text-muted-foreground @max-[560px]:hidden">
-          {streamCount} {streamCount === 1 ? "message" : "messages"}
-        </span>
-      )}
-    </div>
-  );
-};
-
-interface HeadersContentProps {
-  methodCall: MethodCall;
-}
-
-Console.HeadersContent = function ({ methodCall }: HeadersContentProps) {
-  const requestHeaders = methodCall.requestHeaders || {};
-  const responseHeaders = methodCall.responseHeaders || {};
-  const upstreamRequestHeaders = methodCall.upstreamRequestHeaders || {};
-  const upstreamResponseHeaders = methodCall.upstreamResponseHeaders || {};
-  // The request line of the upstream call, which a failure reports and the
-  // response no longer carries. A successful call doesn't report one.
-  const upstreamRequest = upstreamRequestLine(methodCall.error);
-  // An in-process app (e.g. OpenAPI) reports the headers it exchanged with its
-  // upstream API. When present, the transport headers (browser ↔ Kaja) become a
-  // second, less interesting hop shown below the upstream ones.
-  const hasUpstream = upstreamRequest !== undefined || Object.keys(upstreamRequestHeaders).length > 0 || Object.keys(upstreamResponseHeaders).length > 0;
-
-  const section = (title: string, headers: { [key: string]: string }) => (
-    <div className="mb-6">
-      <div className="mb-2 font-semibold text-foreground">{title}</div>
-      {Object.keys(headers).length > 0 ? (
-        <Console.HeadersTable headers={headers} />
-      ) : (
-        <div className="italic text-muted-foreground">No {title.toLowerCase()}</div>
-      )}
-    </div>
-  );
-
-  const groupHeading = (text: string, caption: string, requestLine?: string) => (
-    <div className="mb-3">
-      <div className="font-semibold uppercase tracking-wider text-foreground">{text}</div>
-      <div className="text-muted-foreground">{caption}</div>
-      {requestLine && <div className="mt-1 break-all text-foreground">{requestLine}</div>}
-    </div>
-  );
-
-  return (
-    <div className="min-h-0 flex-1 overflow-auto p-4 font-mono text-xs">
-      {hasUpstream ? (
-        <>
-          {groupHeading("Upstream", "Headers Kaja exchanged with the API", upstreamRequest)}
-          {section("Request headers", upstreamRequestHeaders)}
-          {section("Response headers", upstreamResponseHeaders)}
-          <div className="mb-6 h-px bg-border" />
-          {groupHeading("Transport", "Headers between the browser and Kaja")}
-          {section("Request headers", requestHeaders)}
-          {section("Response headers", responseHeaders)}
-        </>
-      ) : (
-        <>
-          {section("Request Headers", requestHeaders)}
-          {section("Response Headers", responseHeaders)}
-        </>
-      )}
-    </div>
-  );
-};
-
-interface HeadersTableProps {
-  headers: { [key: string]: string };
-}
-
-Console.HeadersTable = function ({ headers }: HeadersTableProps) {
-  const sortedKeys = Object.keys(headers).sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
-
-  return (
-    <table className="w-full border-collapse">
-      <tbody>
-        {sortedKeys.map((key) => (
-          <tr key={key}>
-            <td className="whitespace-nowrap py-1 pr-3 align-top text-muted-foreground">{key}:</td>
-            <td className="break-all py-1 text-foreground">{headers[key]}</td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-};
 
 interface RunSummaryLine {
   name: string;
